backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.1
PyJWT==2.8.0
Werkzeug==3.0.1
